MFC r341068:
Make whitespace more consistent in libkqueue tests.

After r337820, which "corrected" some spaces-instead-of-tab whitespace
issues in the libkqueue tests, jmg@ pointed out that these files were
originally space-based, not tab-spaced, and so the correction should
have been to get rid of the tabs that had been introduced in previous
changes, not the spaces. This change does that. This is a whitespace
only change; no functional change is intended.

Reported by:	jmg@
Sponsored by:	Dell EMC Isilon


git-svn-id: https://svn.freebsd.org/base/stable/12@341274 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f