MFC r340777:
Ensure that the default RTT stack can make an RTT measurement if
the TCP connection was initiated using the RACK stack, but the
peer does not support the TCP RACK extension.

This ensures that the TCP behaviour on the wire is the same if
the TCP connection is initated using the RACK stack or the default
stack.

Reviewed by:		rrs@
Sponsored by:		Netflix, Inc.
Differential Revision:	https://reviews.freebsd.org/D18032


git-svn-id: https://svn.freebsd.org/base/stable/12@341499 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f