MFC r341530: ibcore: Fix loopback with rdma-cm.
Trying to validate loopback fails because rtalloc1() resolves system
local addresses to the loopback network interface, lo0. Fix this by
explicitly checking for loopback during validation of the source
and destination network address. If the source address belongs to
a local network interface and is equal to the destination address,
there is no need to run the destination address through rtalloc1().

Sponsored by:   Mellanox Technologies


git-svn-id: https://svn.freebsd.org/base/stable/11@341874 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f