MFC r341570: mlx5ib: Make sure the congestion work timer does not escape the drain procedure.
If the mlx5_ib_read_cong_stats() function was running when mlx5ib was unloaded,
because this function unconditionally restarts the timer, the timer can still
be pending after the delayed work has been cancelled. To fix this simply loop
on the delayed work cancel procedure as long as it returns non-zero.

Sponsored by:   Mellanox Technologies


git-svn-id: https://svn.freebsd.org/base/stable/11@341954 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f