MFC 339620: Add a "live" mode to ktrdump.
Support a "live" mode in ktrdump enabled via the -l flag.  In this
mode, ktrdump polls the kernel's trace buffer periodically (currently
hardcoded as a 50 millisecond interval) and dumps any newly added
entries.  Fancier logic for the timeout (e.g. a command line option or
some kind of backoff based on the time since the last entry) can be
added later as the need arises.

Note that this does not include the capsicum fixes from the original
commit since ktrdump in stable/11 is not capsicum-ized.


git-svn-id: https://svn.freebsd.org/base/stable/11@342706 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f