MFS12 r342378: Fix a regression in the TCP handling of received segments.
When receiving TCP segments the stack protects itself by limiting
the resources allocated for a TCP connections. This patch adds
an exception to these limitations for the TCP segement which is the next
expected in-sequence segment. Without this patch, TCP connections
may stall and finally fail in some cases of packet loss.

Approved by:	so
Security:	FreeBSD-EN-19:02.tcp


git-svn-id: https://svn.freebsd.org/base/releng/12.0@342894 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f