powerpc/pseries: Cache the IPI vector to avoid the common static lookup
The IPI vector is static, and happens to be the most common interrupt by far
on some systems.  Rather than searching for the interrupt every time, cache
the index.

This appears to yield a small performance boost, of about 8% reduction in
buildworld times, on my POWER9 system, when paired with r342975.


git-svn-id: https://svn.freebsd.org/base/head@342976 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f