pfctl: Fix 'set skip' handling for groups
When we skip on a group the kernel will automatically skip on the member
interfaces. We still need to update our own cache though, or we risk
overruling the kernel afterwards.

This manifested as 'set skip' working initially, then not working when
the rules were reloaded.

PR:		229241
MFC after:	1 week


git-svn-id: https://svn.freebsd.org/base/head@342989 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f