MFC r342966: net80211: fix possible panic for some drivers after r342463
Check if rate control structures were allocated before trying to
access them in various places; this was possible before on
allocation failure (unlikely), but was revealed after r342211
where allocation was deferred.

In case if driver uses wlan_amrr(4) and it is loaded it
is possible to reproduce the panic via

sysctl net.wlan.<number>.rate_stats

(for wlan0 the number will be 0).


git-svn-id: https://svn.freebsd.org/base/stable/12@343033 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f