MFC r342966: net80211: fix possible panic for some drivers after r342464
Check if rate control structures were allocated before trying to
access them in various places; this was possible before on
allocation failure (unlikely), but was revealed after r342211
where allocation was deferred.

In case if driver uses wlan_amrr(4) and it is loaded it
is possible to reproduce the panic via

sysctl net.wlan.<number>.rate_stats

(for wlan0 the number will be 0).

The patch was adjusted a bit since file contents are different enough
since r306591.


git-svn-id: https://svn.freebsd.org/base/stable/11@343035 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f