MFC r342966: net80211: fix possible panic for some drivers after r342465
Check if rate control structures were allocated before trying to
access them in various places; this was possible before on
allocation failure (unlikely), but was revealed after r342211
where allocation was deferred.

The patch was adjusted a bit since file contents are different enough
since r306591.

Also, 'rate_stats' sysctl is not available here, so the panic is
unlikely to happen here even without this patch.


git-svn-id: https://svn.freebsd.org/base/stable/10@343036 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f