MFC r305074-r305075, r327275, r327570: newfs_msdos updates
r305074:
Remove CHS alignment. It's not needed and causes problems for the BBB
boot partition. NetBSD removed it in 1.10 in their repo some time ago.

r305075:
The code only converts from bpbHugeSectors to bpbSectors if the sum of
the hidden and huge sectors is less than or equal MAXU16. When
formatting in Windows bpbSectors is still used for 63488 sectors and
2048 hidden (sum > MAXU16). The hidden sectors count is the number of
sectors before the FAT16 Boot Record so it shouldn't affect the sector
count. Attached patch (huge_sec_conversion.patch) to only check for
bpb.bpbHugeSectors <= MAXU16 when converting to bpbSectors.

r327275:
Close fd and fd1 before returning now that we're done with them.

r327570:
Only call close if fd and fd1 are not -1.

PR: 183234


git-svn-id: https://svn.freebsd.org/base/stable/11@343044 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f