MFC r342874: Fix bsdgrep manpage clobbering grep(1) with default build
The default build should install bsdgrep(1) and grep(1), with the latter
being gnugrep(1). WITH_BSD_GREP flips this situation such that we have
gnugrep(1) and grep(1), with the latter being bsdgrep(1).

Changes to start installing the zgrep script out of usr.bin/grep
inadvertently altered the default build such that grep(1) was being
installed, and it was bsdgrep(1). Correct the typo.


git-svn-id: https://svn.freebsd.org/base/stable/12@343050 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f