MFC 340466,340473: Move the TLS key map into the adapter softc.
340466:
Move the TLS key map into the adapter softc so non-TOE code can use it.

340473:
Restore the <sys/vmem.h> header to fix build of cxgbe(4) TOM.

vmem's are not just used for TLS memory in TOM and the #include actually
predates the TLS code so should not have been removed when the TLS vmem
moved in r340466.


git-svn-id: https://svn.freebsd.org/base/stable/12@343062 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f