Document that `sendfile` will return an invalid value for `sbytes` if provided an invalid address
This is meant to clarify the fact that the system call will not fail
with -1/EFAULT, as one might expect, when reading the sendfile(2)
manpage today.

While here, pet the mandoc linter, when dealing with the section that
describes valid values for `flags`.

PR:	232210
MFC after:	2 weeks
Approved by:	emaste (mentor)
Reviewed by:	glebius, 0mp
Differential Revision: https://reviews.freebsd.org/D18949


git-svn-id: https://svn.freebsd.org/base/head@343444 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f